+++ c/util/simd/org_lwjgl_util_simd_SSE.c
@@ -10,7 +10,7 @@ EXTERN_C_ENTER
 
 JNIEXPORT void JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1SET_1EXCEPTION_1STATE(JNIEnv *__env, jclass clazz, jint mask) {
     UNUSED_PARAMS(__env, clazz)
-    _MM_SET_EXCEPTION_STATE(mask);
+    _MM_SET_EXCEPTION_STATE((unsigned int)mask);
 }
 
 JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1EXCEPTION_1STATE(JNIEnv *__env, jclass clazz) {
@@ -20,7 +20,7 @@ JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1EXCEPTION_1STATE(
 
 JNIEXPORT void JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1SET_1EXCEPTION_1MASK(JNIEnv *__env, jclass clazz, jint mask) {
     UNUSED_PARAMS(__env, clazz)
-    _MM_SET_EXCEPTION_MASK(mask);
+    _MM_SET_EXCEPTION_MASK((unsigned int)mask);
 }
 
 JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1EXCEPTION_1MASK(JNIEnv *__env, jclass clazz) {
@@ -30,7 +30,7 @@ JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1EXCEPTION_1MASK(J
 
 JNIEXPORT void JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1SET_1ROUNDING_1MODE(JNIEnv *__env, jclass clazz, jint mode) {
     UNUSED_PARAMS(__env, clazz)
-    _MM_SET_ROUNDING_MODE(mode);
+    _MM_SET_ROUNDING_MODE((unsigned int)mode);
 }
 
 JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1ROUNDING_1MODE(JNIEnv *__env, jclass clazz) {
@@ -40,7 +40,7 @@ JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1ROUNDING_1MODE(JN
 
 JNIEXPORT void JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1SET_1FLUSH_1ZERO_1MODE(JNIEnv *__env, jclass clazz, jint mode) {
     UNUSED_PARAMS(__env, clazz)
-    _MM_SET_FLUSH_ZERO_MODE(mode);
+    _MM_SET_FLUSH_ZERO_MODE((unsigned int)mode);
 }
 
 JNIEXPORT jint JNICALL Java_org_lwjgl_util_simd_SSE__1MM_1GET_1FLUSH_1ZERO_1MODE(JNIEnv *__env, jclass clazz) {
