+++ c/util/zstd/org_lwjgl_util_zstd_Zdict.c
@@ -16,7 +16,7 @@ JNIEXPORT jlong JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1trainFromBuffer(J
     const void *samplesBuffer = (const void *)(intptr_t)samplesBufferAddress;
     const size_t *samplesSizes = (const size_t *)(intptr_t)samplesSizesAddress;
     UNUSED_PARAMS(__env, clazz)
-    return (jlong)ZDICT_trainFromBuffer(dictBuffer, (size_t)dictBufferCapacity, samplesBuffer, samplesSizes, nbSamples);
+    return (jlong)ZDICT_trainFromBuffer(dictBuffer, (size_t)dictBufferCapacity, samplesBuffer, samplesSizes, (unsigned int)nbSamples);
 }
 
 JNIEXPORT jint JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1getDictID(JNIEnv *__env, jclass clazz, jlong dictBufferAddress, jlong dictSize) {
@@ -41,7 +41,7 @@ JNIEXPORT jlong JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1trainFromBuffer_1
     const size_t *samplesSizes = (const size_t *)(intptr_t)samplesSizesAddress;
     ZDICT_cover_params_t *parameters = (ZDICT_cover_params_t *)(intptr_t)parametersAddress;
     UNUSED_PARAMS(__env, clazz)
-    return (jlong)ZDICT_trainFromBuffer_cover(dictBuffer, (size_t)dictBufferCapacity, samplesBuffer, samplesSizes, nbSamples, *parameters);
+    return (jlong)ZDICT_trainFromBuffer_cover(dictBuffer, (size_t)dictBufferCapacity, samplesBuffer, samplesSizes, (unsigned int)nbSamples, *parameters);
 }
 
 JNIEXPORT jlong JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1optimizeTrainFromBuffer_1cover(JNIEnv *__env, jclass clazz, jlong dictBufferAddress, jlong dictBufferCapacity, jlong samplesBufferAddress, jlong samplesSizesAddress, jint nbSamples, jlong parametersAddress) {
@@ -50,7 +50,7 @@ JNIEXPORT jlong JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1optimizeTrainFrom
     const size_t *samplesSizes = (const size_t *)(intptr_t)samplesSizesAddress;
     ZDICT_cover_params_t *parameters = (ZDICT_cover_params_t *)(intptr_t)parametersAddress;
     UNUSED_PARAMS(__env, clazz)
-    return (jlong)ZDICT_optimizeTrainFromBuffer_cover(dictBuffer, (size_t)dictBufferCapacity, samplesBuffer, samplesSizes, nbSamples, parameters);
+    return (jlong)ZDICT_optimizeTrainFromBuffer_cover(dictBuffer, (size_t)dictBufferCapacity, samplesBuffer, samplesSizes, (unsigned int)nbSamples, parameters);
 }
 
 JNIEXPORT jlong JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1finalizeDictionary(JNIEnv *__env, jclass clazz, jlong dictBufferAddress, jlong dictBufferCapacity, jlong dictContentAddress, jlong dictContentSize, jlong samplesBufferAddress, jlong samplesSizesAddress, jint nbSamples, jlong parametersAddress) {
@@ -60,7 +60,7 @@ JNIEXPORT jlong JNICALL Java_org_lwjgl_util_zstd_Zdict_nZDICT_1finalizeDictionar
     const size_t *samplesSizes = (const size_t *)(intptr_t)samplesSizesAddress;
     ZDICT_params_t *parameters = (ZDICT_params_t *)(intptr_t)parametersAddress;
     UNUSED_PARAMS(__env, clazz)
-    return (jlong)ZDICT_finalizeDictionary(dictBuffer, (size_t)dictBufferCapacity, dictContent, (size_t)dictContentSize, samplesBuffer, samplesSizes, nbSamples, *parameters);
+    return (jlong)ZDICT_finalizeDictionary(dictBuffer, (size_t)dictBufferCapacity, dictContent, (size_t)dictContentSize, samplesBuffer, samplesSizes, (unsigned int)nbSamples, *parameters);
 }
 
 EXTERN_C_EXIT
